Controll axis polarity & events outside container
POLARITY

A polarity stream allows you to switch direction
externally for either axis.  It also allows to to
disable an axis entirely by setting its polarity to 0.

ORIGIN CHECK

Listen to all events on the document, but only respond if they originated (mousedown) in the container.

This allows you to start dragging in a container and exit the container without interrupting the drag.